Improve channels DB migration performance
Migrating the channels table by decoding every channel and re-writing
every row is extremely inefficient. On large databases, it will run into
WAL checkpoint issues.

It is much more efficient to update this table directly in SQL, even
though it doesn't guarantee that all columns will correctly be filled if
the JSON column wasn't up-to-date. We thus fill the new `remote_node_id`
column whenever the channel is updated to ensure that it is eventually
filled.