Fix interpretation of option_onion_messages
According to the spec, option_onion_messages signals that the node can forward onion messages which is different from being able to send or receive them (there is no feature bit for that).
We now allow nodes with thi feature disabled to still receive messages and remove the NoRelay relay policy as it is redundant.