Include splicing in the quiescence timeout
Being quiescent for too long is dangerous, because HTLCs may timeout
without giving the opportunity to our peer to send us a preimage.

Splicing operations shouldn't take long to complete, so we include that
in the quiescence timeout, and disconnect if the splice wasn't completed
in time. This makes sure we replay pending commands and avoid getting
the channel stuck if the interactive-tx isn't making progress.