Remove `handleSendRevocation`
This function was called in only one place, so it's not avoiding code
duplication. It could make sense to isolate the logic, but in the case
of event handlers in the highly critical channel FSM, we usually like to
inline transition and state changes to make it easier to review the code
linearly (no need to jump back and forth between function calls).

I also renamed `isIdle` to `isQuiescent` to remove confusion.