Refactor `handleNewSplice` and a few nits
In channel event handlers, we try to leave all the logic that updates the
current state or data directly in the event handler instead of delegating
it to helper functions, otherwise it's a bit hard to see all the state
transitions that can happen when receiving a given message. I changed the
`handleNewSplice` function to be a pure function instead that just creates
the `splice_init` message (or returns an error).

I also refactored some nits.