Allow splicing on non dual-funded channels (#2727)
* Allow splicing on non dual-funded channels

We currently allow splicing on top of a channel that was created without
using dual funding, but it results in an incorrect channel reserve being
used post-splice.

* Rename requestedChannelReserve_opt

This field only applies to the first funding index, so we rename it to
make that explicit. Once a splice has happened, it will be ignored.