Better feerate estimation without change outputs
When asking bitcoind to fund the transaction, we always include the
shared input and output in the transaction even when we're not
the initiator, and later deduce the corresponding fees from our change
output. But if bitcoind doesn't add a change output, we end up paying
a lot more fees than the targeted feerate.

In the special cases where we're responding to a liquidity ads, we
don't need to include the shared input in the tx construction,
which reduces the potential overpayment.