Omit previous transaction for taproot inputs
We include the whole previous transaction in `tx_add_input` to protect
against malleability attacks that affect segwit v0. Since lightning
messages are limited to 65kB, this has an annoying consequence: if the
input you're trying to spend comes from a large transaction, you cannot
use it with the `interactive-tx` protocol.

This is happening in practice with coinjoins and payouts from mining
pools. Since segwit v1+ is not affected by this malleability issue,
we can relax that requirement and only send the corresponding `txout`.