Work around `fundrawtransaction` consolidation
We don't have any control over bitcoind's coin selection: it sometimes
ends up creating transactions with a large number of inputs (and thus
a high fee), which we want to avoid. We run multiple coin selection
attempts in parallel and use the best result to mitigate that.

We should monitor how this behaves in practice and revert that change
if it doesn't prove to be useful.

We should also push for more configuration options on the bitcoind side
to get funding results that better match our transaction patterns.